samhita-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
sqlalchemy
aiosqlite
psycopg2-binary
PyMuPDF
Pillow
google-genai
groq
sentence-transformers
faiss-cpu
fhir.resources
python-dotenv
supabase
httpx
requests
livekit-api
gunicorn
